% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_and_diagnosis.R
\name{determine_ptsd_diagnosis}
\alias{determine_ptsd_diagnosis}
\title{Determine PTSD diagnosis (non-binarized)}
\usage{
determine_ptsd_diagnosis(data)
}
\arguments{
\item{data}{A dataframe with at least 20 columns
(named symptom_1 to symptom_20), for each symptom a value between 0-4}
}
\value{
A dataframe with an additional column "PTSD_Diagnosis" telling
if the Diagnosis criteria are fulfilled
}
\description{
Determines PTSD diagnosis on the basis of the non-binarized values
of the 20 items
}
