% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_optimizations.R
\name{print.ptsdiag_comparison}
\alias{print.ptsdiag_comparison}
\title{Print method for ptsdiag_comparison objects}
\usage{
\method{print}{ptsdiag_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{ptsdiag_comparison} object.}

\item{...}{Unused.}
}
\value{
The input object invisibly.
}
\description{
Print method for ptsdiag_comparison objects
}
