% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_and_diagnosis.R
\name{summarize_ptsd}
\alias{summarize_ptsd}
\title{Summarize PTSD scores and diagnosis}
\usage{
summarize_ptsd(data)
}
\arguments{
\item{data}{A dataframe containing PTSD total scores and diagnoses}
}
\value{
A dataframe with summary statistics
}
\description{
Creates a summary of PTSD total scores and diagnoses
}
