% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_best_six_symptoms_four_required}
\alias{analyze_best_six_symptoms_four_required}
\title{Find best three six-symptom-combinations for non-hierarchical PTSD diagnosis}
\usage{
analyze_best_six_symptoms_four_required(data, score_by = "false_cases")
}
\arguments{
\item{data}{A dataframe with 20 columns for the 20 symptoms with
non-binarized values}

\item{score_by}{Method to select the best six-symptom-combinations
(by minimizing the "newly_nondiagnosed" or by minimizing the "false_cases")}
}
\value{
A list containing best combinations and diagnosis comparison
}
\description{
Determine the three best six-symptom-combinations, where at least four
symptoms need to be present for diagnosis, regardless of which cluster
they belong to
}
