% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{binarize_data}
\alias{binarize_data}
\title{Binarize data}
\usage{
binarize_data(data)
}
\arguments{
\item{data}{A dataframe with values between 0-4 for each of the 20 symptoms}
}
\value{
A dataframe with 0 (=absent) or 1 (=present) for each of the 20 symptoms
}
\description{
Symptoms with values of 0 or 1 are considered as absent, symptom values of ≥2
are considered as present
}
