% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_and_diagnosis.R
\name{calculate_ptsd_total}
\alias{calculate_ptsd_total}
\title{Calculate PTSD total score}
\usage{
calculate_ptsd_total(data)
}
\arguments{
\item{data}{A dataframe (output of rename_ptsd_columns) containing values
between 0-4 for each PTSD item}
}
\value{
A dataframe with an additional column "total" containing the
total score
}
\description{
Calculates the totals score from PTSD symptoms
}
