% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{create_ptsd_diagnosis_binarized}
\alias{create_ptsd_diagnosis_binarized}
\title{Determine PTSD Diagnosis (binarized)}
\usage{
create_ptsd_diagnosis_binarized(data)
}
\arguments{
\item{data}{A dataframe with raw symptom scores of the 20 symptoms}
}
\value{
A dataframe with one column
indicating whether diagnostic criteria are met
}
\description{
Determines PTSD on the basis of the binarized values for the 20 items
}
