% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{create_readable_summary}
\alias{create_readable_summary}
\title{Create readable summary of changes in diagnostic metrics when comparing
different diagnostic criteria}
\usage{
create_readable_summary(summary_stats, columns = NULL)
}
\arguments{
\item{summary_stats}{Resulting dataframe of function summarize_ptsd_changes}

\item{columns}{Specific columns to include in summary}
}
\value{
A formatted dataframe with summary statistic for the
different diagnostic criteria
}
\description{
Create readable summary of changes in diagnostic metrics when comparing
different diagnostic criteria
}
