% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{rename_ptsd_columns}
\alias{rename_ptsd_columns}
\title{Rename PTSD symptom columns}
\usage{
rename_ptsd_columns(data)
}
\arguments{
\item{data}{A dataframe containing 20 columns, one for each PTSD symptom}
}
\value{
A dataframe with renamed columns
}
\description{
Rename columns in the input data to standardize symptom names
(symptom_1 to symptom_20)
}
