% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping_functions.R
\name{summarize_ptsd_changes}
\alias{summarize_ptsd_changes}
\title{Summarize changes in diagnostic metrics
when comparing different diagnostic criteria}
\usage{
summarize_ptsd_changes(data)
}
\arguments{
\item{data}{A dataframe with columns showing whether the diagnosis is
fulfilled under certain diagnostic criteria}
}
\value{
Summary statistic for the different diagnostic criteria
}
\description{
Summarize changes in diagnostic metrics
when comparing different diagnostic criteria
}
